+++ src/oms_sensemaking/inference/rules/in_out_garrison.py
@@ -14,58 +14,54 @@ from oms_sdk.generated.generated_graphql_client import (
     UuidQueryByList,
 )
 
-from oms_sensemaking.clients.instances import oms_crud_tool
 from oms_sensemaking.config import SETTINGS
 from oms_sensemaking.core.geo_helpers import generate_circle_points_geographical
+from oms_sensemaking.core.oms_crud import OmsCrudTool
+from oms_sensemaking.core.sensemakers import Sensemaker
 from oms_sensemaking.domain.in_or_out_garrison.utils import in_garrison
-from oms_sensemaking.inference.rules.base_rule import BaseRule
 from oms_sensemaking.inference.rules.garrison_data_collection import GetGarrisonDataRetrieverFactory
-from oms_sensemaking.inference.rules.rule_context import RuleContext
 from oms_sensemaking.inference.rules.rule_helper_classes import GeoTimeframe, Timeframe
 
 
-class InOrOutOfGarrison(BaseRule):
+class InOrOutOfGarrison(Sensemaker):
     """
     Detect when a node is in or out of garrison and create/update the appropriate activity
     """
 
-    def __init__(self, name: str):
-        self.name = name
+    def __init__(self, oms_crud_tool: OmsCrudTool):
+        super().__init__()
+        self.oms_crud_tool = oms_crud_tool
+        self.name = self.__class__.__name__
         self.version = (1, 0, 0)
-        self._data_retriever = GetGarrisonDataRetrieverFactory(oms_crud_tool).get_garrison_data_retriever(
+        self._data_retriever = GetGarrisonDataRetrieverFactory(self.oms_crud_tool).get_garrison_data_retriever(
             SETTINGS.garrison_data_retriever
         )
 
-    def evaluate(self, rule_context: RuleContext) -> bool:
+    def evaluate(self, obs: ObservationObservation) -> bool:
         """
         Valid inputs must contain observations that have geometry and point to a node
 
         :param rule_context: Rule context object containing the observation to evaluate
         """
 
-        if (
-            not rule_context.observation
-            or rule_context.observation.startTime is None
-            or rule_context.observation.endTime is None
-        ):
+        if not obs or obs.startTime is None or obs.endTime is None:
             return False
-
-        obs = rule_context.observation
 
         return obs and obs.nodeId and obs.geometry and obs.classIri != SETTINGS.track_iri
 
-    def action(self, rule_context: RuleContext):
+    def process_data(self, obs: ObservationObservation, config: dict | None = None):
         """
         Determine if an observation indicates that a node is in or out of garrison
 
         :param rule_context: Rule context object containing the observation in question
         """
 
-        obs = rule_context.observation
-        node_object = oms_crud_tool.get_node(obs.nodeId)
+        if not self.evaluate(obs) or self.has_action_already_ran(obs):
+            return []
+        node_object = self.oms_crud_tool.get_node(obs.nodeId)
         object_and_garrison_coords = self._data_retriever.get_all_garrison_data(obs)
         if not object_and_garrison_coords:
-            return None
+            return []
         object_lat_lon, garrison_lat_lon = object_and_garrison_coords
         in_garrison_check = in_garrison(object_lat_lon, garrison_lat_lon)
         garrison_buffer_points = generate_circle_points_geographical(
@@ -74,14 +70,7 @@ class InOrOutOfGarrison(BaseRule):
         garrison_buffer_geojson = {"type": "Polygon", "coordinates": [garrison_buffer_points]}
 
         self._create_or_update_garrison_activity(obs, node_object, in_garrison_check, garrison_buffer_geojson)
-
-    def get_graph_data(self, rule_context: RuleContext):
-        """
-        Retrieve all data from Atoms API required for this rule.
-
-        :param rule_context: Rule context object containing the observation in question
-        """
-        return
+        return []
 
     def _create_or_update_garrison_activity(
         self, obs: ObservationObservation, node_object: NodeNode, in_garrison: bool, garrison_buffer_geojson: dict
@@ -101,7 +90,7 @@ class InOrOutOfGarrison(BaseRule):
             states=[activity_state],
             nodeIds=UuidQueryByList(in_=[obs.nodeId]),
         )
-        activity_response = oms_crud_tool.get_activities(activity_query)
+        activity_response = self.oms_crud_tool.get_activities(activity_query)
         existing_activities = activity_response.data
 
         matching_activity_found = False
@@ -143,7 +132,7 @@ class InOrOutOfGarrison(BaseRule):
             observationIds=UpdateUuidList(add=[observation.id]),
             nodeId=observation.nodeId,
         )
-        oms_crud_tool.update_activity(updated_activity_input)
+        self.oms_crud_tool.update_activity(updated_activity_input)
 
     def _handle_new_activity(self, observation: ObservationObservation, activity_name: str, activity_state: str):
         """
@@ -167,19 +156,18 @@ class InOrOutOfGarrison(BaseRule):
             startTime=observation.startTime,
             endTime=observation.endTime,
         )
-        oms_crud_tool.create_activity(garrison_activity)
+        self.oms_crud_tool.create_activity(garrison_activity)
 
-    def has_action_already_ran(self, rule_context: RuleContext):
+    def has_action_already_ran(self, obs: ObservationObservation):
         """
         Determine if an in/out of garrison activity pointing to the inputted observation has already been created
         """
 
-        obs = rule_context.observation
         if not obs:
             return False
 
         activity_query = ActivityQuery(observationIds=[obs.id])
-        activities = oms_crud_tool.get_activities(activity_query).data
+        activities = self.oms_crud_tool.get_activities(activity_query).data
 
         return any(
             act.name == SETTINGS.inference_in_garrison_activity_name